+++ src/main/java/com/example/coffeeproject/Coffee/Model/ComposeCoffee.java
@@ -16,7 +16,7 @@ public class ComposeCoffee {
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
     private int compose_id;
 
-    private int ounce; // 중량 1oz=29.5ml
+    private double ounce; // 중량 1oz=29.5ml
     private double price; // 가격
     private String name; // 이름
     private String eng_name; // 영어 이름
